% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoa.R
\name{demultiplex}
\alias{demultiplex}
\title{Demultiplex nanopore reads}
\usage{
demultiplex(
  samples,
  filenames,
  outdir,
  min_score_start = 10,
  min_score_end = 10,
  better_by = 6,
  min_length = 20
)
}
\description{
Demultiplex nanopore reads
}
