% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoa.R
\name{examine_hits}
\alias{examine_hits}
\title{Examine alignments to a sample of reads}
\usage{
examine_hits(samples, filenames, min_score = 20, n_per_file = 10)
}
\arguments{
\item{samples}{A data frame with three columns. The first column is the sample name (will be used in the output filename). The second column is the left adaptor sequence. The third column is the right adaptor sequence.}

\item{filenames}{A character vector of fastq read filenames.}

\item{min_score}{Minimum alignment score to show.}

\item{n_per_file}{Number of reads to sample per file.}
}
\description{
Sample reads from fastq files, and show where adaptor sequences align to them.
}
